code/requirements.txt
click==8.1.3
click==8.1.7
matplotlib==3.9.2
numpy==2.1.1
opencv_python==4.10.0.84
pandas==2.2.2
Pillow==10.4.0
scikit_learn==1.5.2
seaborn==0.13.2
torch==2.4.1+cu118
torchvision==0.19.1+cu118
tqdm==4.66.4
